fix: jobBoard 카테고리api는 category 도메인으로 변경
